con message vs MVC

+++ admin/View/admin_product.php
@@ -63,126 +63,127 @@ if (isset($_POST['update_product'])) {
     <title>Quản lý thêm, sửa, xóa sách</title>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
     <link rel="stylesheet" href="../../public/css/admin.css">
+    <link rel="stylesheet" href="../../public/css/tuananh.css">
     <style>
-    .blackboard {
-        position: relative;
-        width: 640px;
-        margin: 5% auto;
-        border: tan solid 12px;
-        border-top: #bda27e solid 12px;
-        border-left: #b19876 solid 12px;
-        border-bottom: #c9ad86 solid 12px;
-        box-shadow: 0px 0px 6px 5px rgba(58, 18, 13, 0), 0px 0px 0px 2px #c2a782, 0px 0px 0px 4px #a58e6f, 3px 4px 8px 5px rgba(0, 0, 0, 0.5);
-        background-image: radial-gradient(circle at left 30%, rgba(34, 34, 34, 0.3), rgba(34, 34, 34, 0.3) 80px, rgba(34, 34, 34, 0.5) 100px, rgba(51, 51, 51, 0.5) 160px, rgba(51, 51, 51, 0.5)), linear-gradient(215deg, transparent, transparent 100px, #222 260px, #222 320px, transparent), radial-gradient(circle at right, #111, rgba(51, 51, 51, 1));
-        background-color: #333;
-    }
+        .blackboard {
+            position: relative;
+            width: 640px;
+            margin: 5% auto;
+            border: tan solid 12px;
+            border-top: #bda27e solid 12px;
+            border-left: #b19876 solid 12px;
+            border-bottom: #c9ad86 solid 12px;
+            box-shadow: 0px 0px 6px 5px rgba(58, 18, 13, 0), 0px 0px 0px 2px #c2a782, 0px 0px 0px 4px #a58e6f, 3px 4px 8px 5px rgba(0, 0, 0, 0.5);
+            background-image: radial-gradient(circle at left 30%, rgba(34, 34, 34, 0.3), rgba(34, 34, 34, 0.3) 80px, rgba(34, 34, 34, 0.5) 100px, rgba(51, 51, 51, 0.5) 160px, rgba(51, 51, 51, 0.5)), linear-gradient(215deg, transparent, transparent 100px, #222 260px, #222 320px, transparent), radial-gradient(circle at right, #111, rgba(51, 51, 51, 1));
+            background-color: #333;
+        }
 
-    .blackboard:before {
-        box-sizing: border-box;
-        display: block;
-        position: absolute;
-        width: 100%;
-        height: 100%;
-        background-image: linear-gradient(175deg, transparent, transparent 40px, rgba(120, 120, 120, 0.1) 100px, rgba(120, 120, 120, 0.1) 110px, transparent 220px, transparent), linear-gradient(200deg, transparent 80%, rgba(50, 50, 50, 0.3)), radial-gradient(ellipse at right bottom, transparent, transparent 200px, rgba(80, 80, 80, 0.1) 260px, rgba(80, 80, 80, 0.1) 320px, transparent 400px, transparent);
-        border: #2c2c2c solid 2px;
-        content: "Tìm Kiếm Sản Phẩm";
-        font-family: 'Permanent Marker', cursive;
-        font-size: 2.2em;
-        color: rgba(238, 238, 238, 0.7);
-        text-align: center;
-        padding-top: 20px;
-    }
+        .blackboard:before {
+            box-sizing: border-box;
+            display: block;
+            position: absolute;
+            width: 100%;
+            height: 100%;
+            background-image: linear-gradient(175deg, transparent, transparent 40px, rgba(120, 120, 120, 0.1) 100px, rgba(120, 120, 120, 0.1) 110px, transparent 220px, transparent), linear-gradient(200deg, transparent 80%, rgba(50, 50, 50, 0.3)), radial-gradient(ellipse at right bottom, transparent, transparent 200px, rgba(80, 80, 80, 0.1) 260px, rgba(80, 80, 80, 0.1) 320px, transparent 400px, transparent);
+            border: #2c2c2c solid 2px;
+            content: "Tìm Kiếm Sản Phẩm";
+            font-family: 'Permanent Marker', cursive;
+            font-size: 2.2em;
+            color: rgba(238, 238, 238, 0.7);
+            text-align: center;
+            padding-top: 20px;
+        }
 
-    .form {
-        padding: 70px 20px 20px;
-    }
+        .form {
+            padding: 70px 20px 20px;
+        }
 
-    p {
-        position: relative;
-        margin-bottom: 1em;
-    }
+        p {
+            position: relative;
+            margin-bottom: 1em;
+        }
 
-    label {
-        vertical-align: middle;
+        label {
+            vertical-align: middle;
 
-        font-size: 1.6em;
-        color: rgba(238, 238, 238, 0.7);
-    }
+            font-size: 1.6em;
+            color: rgba(238, 238, 238, 0.7);
+        }
 
-    p:nth-of-type(5)>label {
-        vertical-align: top;
-    }
+        p:nth-of-type(5)>label {
+            vertical-align: top;
+        }
 
-    input,
-    textarea {
-        vertical-align: middle;
-        padding-left: 10px;
-        background: none;
-        border: none;
+        input,
+        textarea {
+            vertical-align: middle;
+            padding-left: 10px;
+            background: none;
+            border: none;
 
-        font-size: 1.6em;
-        color: rgba(238, 238, 238, 0.8);
-        line-height: .6em;
-        outline: none;
-    }
+            font-size: 1.6em;
+            color: rgba(238, 238, 238, 0.8);
+            line-height: .6em;
+            outline: none;
+        }
 
 
-    select {
-        vertical-align: middle;
-        padding-left: 10px;
-        background: transparent;
-        border: none;
-        width: 40%;
+        select {
+            vertical-align: middle;
+            padding-left: 10px;
+            background: transparent;
+            border: none;
+            width: 40%;
 
-        font-size: 1.6em;
-        color: rgba(238, 238, 238, 0.8);
-        line-height: .6em;
-        outline: none;
-    }
+            font-size: 1.6em;
+            color: rgba(238, 238, 238, 0.8);
+            line-height: .6em;
+            outline: none;
+        }
 
-    option {
-        vertical-align: middle;
-        padding-left: 10px;
-        background: rgb(63, 62, 70);
-        border: none;
+        option {
+            vertical-align: middle;
+            padding-left: 10px;
+            background: rgb(63, 62, 70);
+            border: none;
 
-        font-size: 1.0em;
-        color: rgba(238, 238, 238, 0.8);
-        line-height: .6em;
-        outline: none;
-    }
+            font-size: 1.0em;
+            color: rgba(238, 238, 238, 0.8);
+            line-height: .6em;
+            outline: none;
+        }
 
-    textarea {
-        margin-top: 1%;
-        height: 120px;
-        font-size: 1.4em;
-        line-height: 1em;
-        resize: none;
-    }
+        textarea {
+            margin-top: 1%;
+            height: 120px;
+            font-size: 1.4em;
+            line-height: 1em;
+            resize: none;
+        }
 
-    input[type="submit"] {
-        cursor: pointer;
-        color: rgba(238, 238, 238, 0.7);
-        line-height: 1em;
-        padding: 0;
-    }
+        input[type="submit"] {
+            cursor: pointer;
+            color: rgba(238, 238, 238, 0.7);
+            line-height: 1em;
+            padding: 0;
+        }
 
-    input[type="submit"]:focus {
-        background: rgba(238, 238, 238, 0.2);
-        color: rgba(238, 238, 238, 0.2);
-    }
+        input[type="submit"]:focus {
+            background: rgba(238, 238, 238, 0.2);
+            color: rgba(238, 238, 238, 0.2);
+        }
 
-    ::-moz-selection {
-        background: rgba(238, 238, 238, 0.2);
-        color: rgba(238, 238, 238, 0.2);
-        text-shadow: none;
-    }
+        ::-moz-selection {
+            background: rgba(238, 238, 238, 0.2);
+            color: rgba(238, 238, 238, 0.2);
+            text-shadow: none;
+        }
 
-    ::selection {
+        /* ::selection {
         background: rgba(238, 238, 238, 0.4);
         color: rgba(238, 238, 238, 0.3);
         text-shadow: none;
-    }
+    } */
     </style>
 </head>
 
@@ -202,10 +203,8 @@ if (isset($_POST['update_product'])) {
             <input type="text" name="image" class="box" placeholder="Nhập url ảnh" required>
             <input type="text" name="supplier" class="box" placeholder="Nhập nhà cung cấp" required>
             <input type="text" name="publiser" class="box" placeholder="Nhập nhà xuất bản" required>
-            <textarea name="description" class="description" placeholder="Nhập mô tả về sách" cols="30"
-                rows="5"></textarea>
-            <input style="width:180px;height:50px;font-size:23px;" type="submit" value="Thêm Sản Phẩm"
-                name="add_product" class="btn">
+            <textarea name="description" class="description" placeholder="Nhập mô tả về sách" cols="30" rows="5"></textarea>
+            <input style="width:180px;height:50px;font-size:23px;" type="submit" value="Thêm Sản Phẩm" name="add_product" class="btn">
         </form>
 
     </section>
@@ -273,27 +272,24 @@ if (isset($_POST['update_product'])) {
                 if ($sql_products && $sql_products->num_rows > 0) {
                     while ($fetch_products_sql = mysqli_fetch_assoc($sql_products)) {
             ?>
-            <div class="box">
-                <img src="<?php echo $fetch_products_sql['image']; ?>" alt="">
-                <div class="name">
-                    <?php echo $fetch_products_sql['name']; ?>
-                </div>
-                <div class="author">
-                    <?php echo $fetch_products_sql['author']; ?>
-                </div>
-                <div class="price">
-                    <?php echo $fetch_products_sql['price']; ?>
-                    <span class="rate">₫</span></h3>
-                </div>
-                <a href="./admin_view_detail.php?id=<?php echo $fetch_products_sql['product_id'] ?>"
-                    class="detail_book">Xem
-                    thêm <i class="fas fa-angle-right"></i></a> <br>
-                <a href="admin_product.php?update=<?php echo $fetch_products_sql['product_id']; ?>"
-                    class="option-btn">Cập
-                    nhật</a>
-                <a href="admin_product.php?delete=<?php echo $fetch_products_sql['product_id']; ?>" class="delete-btn"
-                    onclick="return confirm('Xóa quyển sách này?');">Xóa</a>
-            </div>
+                        <div class="box">
+                            <img src="<?php echo $fetch_products_sql['image']; ?>" alt="">
+                            <div class="name">
+                                <?php echo $fetch_products_sql['name']; ?>
+                            </div>
+                            <div class="author">
+                                <?php echo $fetch_products_sql['author']; ?>
+                            </div>
+                            <div class="price">
+                                <?php echo $fetch_products_sql['price']; ?>
+                                <span class="rate">₫</span></h3>
+                            </div>
+                            <a href="./admin_view_detail.php?id=<?php echo $fetch_products_sql['product_id'] ?>" class="detail_book">Xem
+                                thêm <i class="fas fa-angle-right"></i></a> <br>
+                            <a href="admin_product.php?update=<?php echo $fetch_products_sql['product_id']; ?>" class="option-btn">Cập
+                                nhật</a>
+                            <a href="admin_product.php?delete=<?php echo $fetch_products_sql['product_id']; ?>" class="delete-btn" onclick="return confirm('Xóa quyển sách này?');">Xóa</a>
+                        </div>
             <?php
                     }
                 } else {
@@ -314,25 +310,24 @@ if (isset($_POST['update_product'])) {
             if (mysqli_num_rows($select_products) > 0) {
                 while ($fetch_products = mysqli_fetch_assoc($select_products)) {
             ?>
-            <div class="box">
-                <img src="<?php echo $fetch_products['image']; ?>" alt="">
-                <div class="name">
-                    <?php echo $fetch_products['name']; ?>
-                </div>
-                <div class="author">
-                    <?php echo $fetch_products['author']; ?>
-                </div>
-                <div class="price">
-                    <?php echo $fetch_products['price']; ?>
-                    <span class="rate">₫</span></h3>
-                </div>
-                <a href="./admin_view_detail.php?id=<?php echo $fetch_products['product_id'] ?>" class="detail_book">Xem
-                    thêm <i class="fas fa-angle-right"></i></a> <br>
-                <a href="admin_product.php?update=<?php echo $fetch_products['product_id']; ?>" class="option-btn">Cập
-                    nhật</a>
-                <a href="admin_product.php?delete=<?php echo $fetch_products['product_id']; ?>" class="delete-btn"
-                    onclick="return confirm('Xóa quyển sách này?');">Xóa</a>
-            </div>
+                    <div class="box">
+                        <img src="<?php echo $fetch_products['image']; ?>" alt="">
+                        <div class="name">
+                            <?php echo $fetch_products['name']; ?>
+                        </div>
+                        <div class="author">
+                            <?php echo $fetch_products['author']; ?>
+                        </div>
+                        <div class="price">
+                            <?php echo $fetch_products['price']; ?>
+                            <span class="rate">₫</span></h3>
+                        </div>
+                        <a href="./admin_view_detail.php?id=<?php echo $fetch_products['product_id'] ?>" class="detail_book">Xem
+                            thêm <i class="fas fa-angle-right"></i></a> <br>
+                        <a href="admin_product.php?update=<?php echo $fetch_products['product_id']; ?>" class="option-btn">Cập
+                            nhật</a>
+                        <a href="admin_product.php?delete=<?php echo $fetch_products['product_id']; ?>" class="delete-btn" onclick="return confirm('Xóa quyển sách này?');">Xóa</a>
+                    </div>
             <?php
                 }
             } else {
@@ -351,26 +346,19 @@ if (isset($_POST['update_product'])) {
             if (mysqli_num_rows($update_query) > 0) {
                 while ($fetch_update = mysqli_fetch_assoc($update_query)) {
         ?>
-        <form action="" method="post" enctype="multipart/form-data">
-            <input type="hidden" name="update_p_id" value="<?php echo $fetch_update['product_id']; ?>">
-            <input type="text" name="update_name" value="<?php echo $fetch_update['name']; ?>" class="box" required
-                placeholder="Nhập tên sách cần cập nhật">
-            <input type="text" name="update_author" value="<?php echo $fetch_update['author']; ?>" class="box" required
-                placeholder="Nhập tên tác gải cần cập nhật">
-            <input type="number" name="update_price" value="<?php echo $fetch_update['price']; ?>" min="0" class="box"
-                required placeholder="Nhập giá sách cần cập nhật">
-            <input type="text" class="box" name="update_image" value="<?php echo $fetch_update['image']; ?>"
-                placeholder="Nhập url ảnh sách cần cập nhật">
-            <!-- <input type="text" class="box" name="update_description" value="" placeholder="Nhập mô tả sách cần cập nhật"> -->
-            <textarea name="update_description" class="box" c ols="30" rows="10"
-                style="width: 100%"><?php echo $fetch_update['description']; ?></textarea>
-            <input type="text" class="box" name="update_supplier" value="<?php echo $fetch_update['supplier']; ?>"
-                placeholder="Nhập nhà cung cấp sách cần cập nhật">
-            <input type="text" class="box" name="update_publiser" value="<?php echo $fetch_update['publiser']; ?>"
-                placeholder="Nhập nhà cung cấp sách cần cập nhật">
-            <input type="submit" value="Lưu" name="update_product" class="btn">
-            <input type="reset" value="Reset" id="close-update" class="option-btn">
-        </form>
+                    <form action="" method="post" enctype="multipart/form-data">
+                        <input type="hidden" name="update_p_id" value="<?php echo $fetch_update['product_id']; ?>">
+                        <input type="text" name="update_name" value="<?php echo $fetch_update['name']; ?>" class="box" required placeholder="Nhập tên sách cần cập nhật">
+                        <input type="text" name="update_author" value="<?php echo $fetch_update['author']; ?>" class="box" required placeholder="Nhập tên tác gải cần cập nhật">
+                        <input type="number" name="update_price" value="<?php echo $fetch_update['price']; ?>" min="0" class="box" required placeholder="Nhập giá sách cần cập nhật">
+                        <input type="text" class="box" name="update_image" value="<?php echo $fetch_update['image']; ?>" placeholder="Nhập url ảnh sách cần cập nhật">
+                        <!-- <input type="text" class="box" name="update_description" value="" placeholder="Nhập mô tả sách cần cập nhật"> -->
+                        <textarea name="update_description" class="box" c ols="30" rows="10" style="width: 100%"><?php echo $fetch_update['description']; ?></textarea>
+                        <input type="text" class="box" name="update_supplier" value="<?php echo $fetch_update['supplier']; ?>" placeholder="Nhập nhà cung cấp sách cần cập nhật">
+                        <input type="text" class="box" name="update_publiser" value="<?php echo $fetch_update['publiser']; ?>" placeholder="Nhập nhà cung cấp sách cần cập nhật">
+                        <input type="submit" value="Lưu" name="update_product" class="btn">
+                        <input type="reset" value="Reset" id="close-update" class="option-btn">
+                    </form>
         <?php
                 }
             }
